Renaming directory and file names for fields
