+++ packages/js/product-editor/src/components/details-categories-field/category-field-add-new-item.tsx
