+++ packages/js/product-editor/src/components/details-categories-field/category-field-item.tsx
