+++ packages/js/product-editor/src/components/details-categories-field/create-category-modal.tsx
