+++ packages/js/product-editor/src/components/details-categories-field/details-categories-field.tsx
