+++ packages/js/product-editor/src/components/details-categories-field/index.ts
@@ -1,0 +1,2 @@
+export * from './details-categories-field';
+export * from './category-field';
