+++ packages/js/product-editor/src/components/details-categories-field/style.scss
