+++ packages/js/product-editor/src/components/details-categories-field/test/category-field.test.tsx
