+++ packages/js/product-editor/src/components/details-categories-field/test/use-category-search.test.ts
