+++ packages/js/product-editor/src/components/details-categories-field/use-category-search.ts
