+++ packages/js/product-editor/src/components/details-description-field/details-description-field.tsx
