+++ packages/js/product-editor/src/components/details-description-field/index.ts
@@ -1,0 +1,1 @@
+export * from './details-description-field';
