+++ packages/js/product-editor/src/components/details-feature-field/details-feature-field.tsx
