+++ packages/js/product-editor/src/components/details-feature-field/index.ts
@@ -1,0 +1,1 @@
+export * from './details-feature-field';
