+++ packages/js/product-editor/src/components/details-name-field/details-name-field.tsx
