+++ packages/js/product-editor/src/components/details-name-field/index.ts
@@ -1,0 +1,1 @@
+export * from './details-name-field';
