+++ packages/js/product-editor/src/components/details-summary-field/details-summary-field.tsx
