+++ packages/js/product-editor/src/components/details-summary-field/index.ts
@@ -1,0 +1,1 @@
+export * from './details-summary-field';
