--- packages/js/product-editor/src/components/field-details-categories/index.ts
@@ -1,2 +1,0 @@
-export * from './field-details-categories';
-export * from './category-field';
