--- packages/js/product-editor/src/components/field-details-description/index.ts
@@ -1,1 +1,0 @@
-export * from './field-details-description';
