--- packages/js/product-editor/src/components/field-details-feature/index.ts
@@ -1,1 +1,0 @@
-export * from './field-details-feature';
