--- packages/js/product-editor/src/components/field-details-name/index.ts
@@ -1,1 +1,0 @@
-export * from './field-details-name';
