+++ packages/js/product-editor/src/components/index.ts
@@ -5,11 +5,8 @@ export {
 export { WooProductFieldItem as __experimentalWooProductFieldItem } from './woo-product-field-item';
 export { WooProductSectionItem as __experimentalWooProductSectionItem } from './woo-product-section-item';
 export { WooProductTabItem as __experimentalWooProductTabItem } from './woo-product-tab-item';
-export { DetailsNameField as __experimentalDetailsNameField } from './field-details-name';
-export { DetailsFeatureField as __experimentalDetailsFeatureField } from './field-details-feature';
-export {
-	DetailsCategoriesField as __experimentalDetailsCategoriesField,
-	CategoryField as __experimentalDetailsCategoryField,
-} from './field-details-categories';
-export { DetailsSummaryField as __experimentalDetailsSummaryField } from './field-details-summary';
-export { DetailsDescriptionField as __experimentalDetailsDescriptionField } from './field-details-description';
+export { DetailsNameField as __experimentalDetailsNameField } from './details-name-field';
+export { DetailsFeatureField as __experimentalDetailsFeatureField } from './details-feature-field';
+export { DetailsCategoriesField as __experimentalDetailsCategoriesField } from './details-categories-field';
+export { DetailsSummaryField as __experimentalDetailsSummaryField } from './details-summary-field';
+export { DetailsDescriptionField as __experimentalDetailsDescriptionField } from './details-description-field';
