+++ packages/js/product-editor/src/style.scss
@@ -1,4 +1,4 @@
 @import 'components/product-section-layout/style.scss';
 @import 'components/edit-product-link-modal/style.scss';
-@import 'components/field-details-categories/style.scss';
-@import 'components/field-details-categories/create-category-modal.scss';
+@import 'components/details-categories-field/style.scss';
+@import 'components/details-categories-field/create-category-modal.scss';
